director/requirements.txt
crispy-forms-bulma==1.1.4
Django==2.2.2
django-avatar==4.1.0
django-allauth==0.39.1
django-configurations==2.1
django-crispy-forms==1.7.2
django-intercom==0.1.2
django-polymorphic==2.0.3
django-rest-auth==0.9.5
django-sendgrid-v5==0.8.0
django-storages==1.7.1
django-debug-toolbar==1.11
django-extensions==2.1.9
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
dj_database_url==0.5.0
google-api-python-client==1.7.9
google-cloud-storage==1.16.1
Pillow==6.0.0
pyjwt==1.7.1
psycopg2-binary==2.8.3
raven==6.10.0
requests==2.22.0
sh==1.12.14
PyGithub==1.43.7
oauth2client==4.1.3